fichier csv

+++ app/Http/Controllers/Admin/ImportFilesController.php
@@ -40,35 +40,48 @@ class ImportFilesController extends Controller
     {
 
         // 1. Validation du fichier uploadé. Extension ".xlsx" autorisée
-        $this->validate($request, [
+        $validation = $this->validate($request, [
             'fichierTD' => 'bail|required|file|mimes:xlsx'
         ]);
 
-        // 2. On déplace le fichier uploadé vers le dossier "public" pour le lire
-        $fichier = $request->fichierTD->move(public_path(), $request->fichierTD->hashName());
+       if ($validation) 
+       {
+           $tdpcu = TDPCU::truncate();
+           if ($tdpcu) 
+           {
+                // 2. On déplace le fichier uploadé vers le dossier "public" pour le lire
+                $fichier = $request->fichierTD->move(public_path(), $request->fichierTD->hashName());
 
-        // 3. $reader : L'instance Spatie\SimpleExcel\SimpleExcelReader
-        $reader = SimpleExcelReader::create($fichier);
- 
-        // On récupère le contenu (les lignes) du fichier
-        $rows = $reader->getRows();
+                // 3. $reader : L'instance Spatie\SimpleExcel\SimpleExcelReader
+                $reader = SimpleExcelReader::create($fichier);
+        
+                // On récupère le contenu (les lignes) du fichier
+                $rows = $reader->getRows();
 
-        // $rows est une Illuminate\Support\LazyCollection
+                // $rows est une Illuminate\Support\LazyCollection
 
-        // 4. On insère toutes les lignes dans la base de données
-        $status = TDPCU::insert($rows->toArray());
+                // 4. On insère toutes les lignes dans la base de données
+                $status = TDPCU::insert($rows->toArray());
 
-        // Si toutes les lignes sont insérées
-        if ($status) {
+                // Si toutes les lignes sont insérées
+                if ($status) {
 
-            // 5. On supprimer le fichier uploadé
-            $reader->close(); // On ferme le $reader
-            unlink($fichier);
+                    // 5. On supprimer le fichier uploadé
+                    $reader->close(); // On ferme le $reader
+                    unlink($fichier);
 
-            // 6. Retour vers le formulaire avec un message $msg
-            return back()->withMsg("Importation réussie !");
+                    // 6. Retour vers le formulaire avec un message $msg
+                    return back()->withMsg("Importation réussie !");
 
-        } else { abort(500); }
+                } else { abort(500); }
+           }
+           else {
+               return back();
+           }
+       }
+       else {
+           return back();
+       }
     
     }
 
@@ -76,36 +89,46 @@ class ImportFilesController extends Controller
     {
 
         // 1. Validation du fichier uploadé. Extension ".xlsx" autorisée
-        $this->validate($request, [
+        $validationDCE =  $this->validate($request, [
             'fichierDCE' => 'bail|required|file|mimes:xlsx'
         ]);
 
-        // 2. On déplace le fichier uploadé vers le dossier "public" pour le lire
-        $fichier = $request->fichierDCE->move(public_path(), $request->fichierDCE->hashName());
+        if ($validationDCE) {
+            $dcepcu = DCEPCU::truncate();
+            if ($dcepcu) {
+                 // 2. On déplace le fichier uploadé vers le dossier "public" pour le lire
+                $fichier = $request->fichierDCE->move(public_path(), $request->fichierDCE->hashName());
 
-        // 3. $reader : L'instance Spatie\SimpleExcel\SimpleExcelReader
-        $reader = SimpleExcelReader::create($fichier);
- 
-        // On récupère le contenu (les lignes) du fichier
-        $rows = $reader->getRows();
+                // 3. $reader : L'instance Spatie\SimpleExcel\SimpleExcelReader
+                $reader = SimpleExcelReader::create($fichier);
+        
+                // On récupère le contenu (les lignes) du fichier
+                $rows = $reader->getRows();
 
-        // $rows est une Illuminate\Support\LazyCollection
+                // $rows est une Illuminate\Support\LazyCollection
 
-        // 4. On insère toutes les lignes dans la base de données
-        $status = DCEPCU::insert($rows->toArray());
+                // 4. On insère toutes les lignes dans la base de données
+                $status = DCEPCU::insert($rows->toArray());
 
-        // Si toutes les lignes sont insérées
-        if ($status) {
+                // Si toutes les lignes sont insérées
+                if ($status) {
 
-            // 5. On supprimer le fichier uploadé
-            $reader->close(); // On ferme le $reader
-            unlink($fichier);
+                    // 5. On supprimer le fichier uploadé
+                    $reader->close(); // On ferme le $reader
+                    unlink($fichier);
 
-            // 6. Retour vers le formulaire avec un message $msg
-            return back()->withMsg("Importation réussie !");
+                    // 6. Retour vers le formulaire avec un message $msg
+                    return back()->withMsg("Importation réussie !");
 
-        } else { abort(500); }
-    
+                } else { abort(500); }
+            }
+            else {
+                return back();
+            }
+        }    
+        else {
+            return back();
+        }
     }
     /**
      * Display the specified resource.
